Fixing 64-bit build warnings in R2_14
--HG--
branch : R2_14

+++ Src/Merge.cpp
@@ -504,7 +504,7 @@ int CMergeApp::DoMessageBox( LPCTSTR lpszPrompt, UINT nType, UINT nIDPrompt )
 		nIDPrompt);
 	
 	// Display the message box dialog and return the result.
-	return dlgMessage.DoModal();
+	return (int) dlgMessage.DoModal();
 }
 
 /**
